+++ src/Rocks.Analysis/Builders/Make/MockConstructorBuilder.cs
@@ -10,8 +10,9 @@ namespace Rocks.Analysis.Builders.Make;
 internal static class MockConstructorBuilder
 {
 	internal static void Build(IndentedTextWriter writer, TypeMockModel mockType, 
-		ImmutableArray<ParameterModel> parameters)
+		ConstructorModel? constructor)
 	{
+		var parameters = constructor is not null ? constructor.Parameters : [];
 		var namingContext = new VariablesNamingContext(parameters);
 		var requiredInitPropertiesAndIndexers = mockType.ConstructorProperties;
 		var hasRequiredProperties = requiredInitPropertiesAndIndexers.Any(_ => _.IsRequired);
@@ -33,7 +34,7 @@ internal static class MockConstructorBuilder
 					return $"{direction}{(_.IsParams ? "params " : string.Empty)}{_.Type.FullyQualifiedName}{requiresNullable} @{_.Name}";
 				})));
 
-		if (hasRequiredProperties)
+		if ((constructor?.RequiresSetsRequiredMembersAttribute ?? false) || hasRequiredProperties)
 		{
 			writer.WriteLine("[global::System.Diagnostics.CodeAnalysis.SetsRequiredMembers]");
 		}