+++ src/Rocks.Analysis/Models/ConstructorModel.cs
@@ -1,11 +1,20 @@
 ﻿using Microsoft.CodeAnalysis;
+using System.Diagnostics.CodeAnalysis;
 
 namespace Rocks.Analysis.Models;
 
 internal sealed record ConstructorModel
 {
-   internal ConstructorModel(IMethodSymbol constructor, ModelContext modelContext) =>
-		this.Parameters = [..constructor.Parameters.Select(_ => new ParameterModel(_, modelContext))];
+	internal ConstructorModel(IMethodSymbol constructor, ModelContext modelContext)
+	{
+		var compilation = modelContext.SemanticModel.Compilation;
+		var setsRequiredMembersAttribute = compilation.GetTypeByMetadataName(typeof(SetsRequiredMembersAttribute).FullName)!;
 
-   internal EquatableArray<ParameterModel> Parameters { get; }
+		this.RequiresSetsRequiredMembersAttribute = constructor.GetAttributes().Any(
+			_ => _.AttributeClass!.Equals(setsRequiredMembersAttribute, SymbolEqualityComparer.Default));
+		this.Parameters = [.. constructor.Parameters.Select(_ => new ParameterModel(_, modelContext))];
+	}
+
+	internal EquatableArray<ParameterModel> Parameters { get; }
+	internal bool RequiresSetsRequiredMembersAttribute { get; }
 }