+++ src/Rocks.Analysis/Polyfills/SetsRequiredMembersAttribute.cs
@@ -1,0 +1,17 @@
+﻿// Lifted from:
+// https://source.dot.net/#System.Private.CoreLib/src/libraries/System.Private.CoreLib/src/System/Diagnostics/CodeAnalysis/SetsRequiredMembersAttribute.cs,e964aa6e099f6872,references
+// Polyfill is needed because this attribute was introduced to support
+// the required keyword in C# in a version where it's not available in NS 2.0
+
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace System.Diagnostics.CodeAnalysis;
+
+/// <summary>
+/// Specifies that this constructor sets all required members for the current type, and callers
+/// do not need to set any required members themselves.
+/// </summary>
+[AttributeUsage(AttributeTargets.Constructor, AllowMultiple = false, Inherited = false)]
+internal sealed class SetsRequiredMembersAttribute 
+	: Attribute { }