+++ changelog.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Interfaces with non-public members that have a default implementation are no longer considered for mocking (issue [#364](https://github.com/JasonBock/Rocks/issues/364))
+- Generated constructors calling a base constructor attributed with `[SetsRequiredMembers]` will now have that attribute as well (issue [#368](https://github.com/JasonBock/Rocks/issues/368))
 
 ## [9.0.2] - 2025-03-17